+++ R/penalties.R
@@ -28,6 +28,11 @@ NULL
 #'     conservation problem to favor solutions that select
 #'     planning units with high connectivity between them.}
 #'
+#'   \item{\code{\link{add_linear_penalties}}}{Add penalties to a
+#'     conservation problem to favor solutions that avoid selecting
+#'     planning units based on a certain variable
+#'     (e.g. anthropogenic pressure).}
+#'
 #'   }
 #'
 #' @seealso \code{\link{constraints}}, \code{\link{decisions}},
@@ -36,7 +41,7 @@ NULL
 #'
 #' @examples
 #' # load data
-#' data(sim_pu_points, sim_features)
+#' data(sim_pu_raster, sim_features)
 #'
 #' # create basic problem
 #' p1 <- problem(sim_pu_raster, sim_features) %>%
@@ -57,14 +62,27 @@ NULL
 #'
 #' # create problem with connectivity penalties
 #' p3 <- p1 %>% add_connectivity_penalties(25, data = scm)
+#'
+#' # create problem with linear penalties,
+#' # here the penalties will be based on random numbers to keep it simple
+#'
+#' # simulate penalty data
+#' sim_penalty_raster <- simulate_cost(sim_pu_raster)
+#'
+#' # plot penalty data
+#' plot(sim_penalty_raster, main = "penalty data", axes = FALSE, box = FALSE)
+#'
+#' # create problem with linear penalties, with a penalty scaling factor of 100
+#' p4 <- p1 %>% add_linear_penalties(100, data = sim_penalty_raster)
+#'
 #' \donttest{
 #' # solve problems
-#' s <- stack(solve(p1), solve(p2), solve(p3))
+#' s <- stack(solve(p1), solve(p2), solve(p3), solve(p4))
 #'
 #' # plot solutions
 #' plot(s, axes = FALSE, box = FALSE,
 #'      main = c("basic solution", "boundary penalties",
-#'               "connectivity penalties"))
+#'               "connectivity penalties", "linear penalties"))
 #'  }
 #' @name penalties
 NULL